refactor(tag): Move ObserveAllTagsSortedUseCase to :core:item
