+++ core/item/src/main/kotlin/de/davis/keygo/core/item/domain/usecase/ObserveAllTagsSortedUseCase.kt
@@ -1,4 +1,4 @@
-package de.davis.keygo.feature.list_screen.domain.usecase
+package de.davis.keygo.core.item.domain.usecase
 
 import de.davis.keygo.core.item.domain.model.Tag
 import de.davis.keygo.core.item.domain.repository.ItemRepository
@@ -8,10 +8,10 @@ import kotlinx.coroutines.flow.map
 import org.koin.core.annotation.Single
 
 @Single
-class ObserveTagUseCase(
+class ObserveAllTagsSortedUseCase(
     private val itemRepository: ItemRepository,
     private val sortUseCase: SortUseCase,
 ) {
     operator fun invoke(): Flow<List<Tag>> =
-        itemRepository.observeAllTags().map { tags -> sortUseCase(tags) }
+        itemRepository.observeAllTags().map { tags -> sortUseCase(tags) { it.display } }
 }
