+++ core/item/src/test/kotlin/de/davis/keygo/core/item/domain/usecase/ObserveAllTagsSortedUseCaseTest.kt
@@ -1,4 +1,4 @@
-package de.davis.keygo.feature.list_screen.domain.usecase
+package de.davis.keygo.core.item.domain.usecase
 
 import de.davis.keygo.core.item.FakeItemRepository
 import de.davis.keygo.core.item.FakeLoginRepository
@@ -12,14 +12,16 @@ import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class ObserveTagUseCaseTest {
+class ObserveAllTagsSortedUseCaseTest {
 
     private val loginRepository = FakeLoginRepository()
     private val itemRepository = FakeItemRepository(loginRepository)
-    private val useCase = ObserveTagUseCase(
+    private val useCase = ObserveAllTagsSortedUseCase(
         itemRepository = itemRepository,
         sortUseCase = SortUseCase(),
     )
+
+    private fun tag(value: String): Tag = Tag.of(value)!!
 
     private fun login(name: String, tags: Set<Tag>) = Login(
         username = null,
@@ -37,13 +39,13 @@ class ObserveTagUseCaseTest {
     @Test
     fun `emits all distinct tags sorted naturally and case-insensitively`() = runTest {
         loginRepository.seed(
-            login("a", setOf("Zebra", "apple")),
-            login("b", setOf("item10", "item2")),
+            login("a", setOf(tag("Zebra"), tag("apple"))),
+            login("b", setOf(tag("item10"), tag("item2"))),
         )
 
         val result = useCase().first()
 
-        assertEquals(listOf("apple", "item2", "item10", "Zebra"), result)
+        assertEquals(listOf(tag("apple"), tag("item2"), tag("item10"), tag("Zebra")), result)
     }
 
     @Test
